Let filepool use full path
A filepool is defined as the full path to the directory (DIRECTORY)
where the files are located including the filename pattern (PATTERN)
that is common for all files. It does not include the file format
(FILE_FORMAT).

```
FILEPOOL=DIRECTORY/PATTERN
```

The files in a pool end with a number and the file ending
specifies the file format of the pool.

file naming must follow

```
FILEPOOL00001.FILE_FORMAT
```